MFC r332773: Fix ddb rc script
r288291 added a call to limits(1), which isn't available before partitions
are mounted. This broke the ddb rc script, which does not provide its own
start_cmd.

Alleviate the situation here by providing a start_cmd. We still have other
problems with diskless setups that need to be considered, but this is a
start.

PR:		206291


git-svn-id: https://svn.freebsd.org/base/stable/11@332864 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f