MFC r332518, r332527
r332518:
Add an option to daemon(8) to specify a delay between restarts of a
supervised program.  The existing -r option has a hard-coded delay of one
second.  This change adds a -R option which takes a delay in seconds.  This
can be used to prevent log spam and rapid restarts, similar to init(8)'s
behavior of adding a delay between rapid restarts when it's supervising a
program.

r332527:
Fix cut-and-pasted line to have the right option letter.


git-svn-id: https://svn.freebsd.org/base/stable/11@332943 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f