Increase the fdtmemreserv array limit to boot on POWER9
Discussing with others, this needs to be at least 20 to boot on some POWER9
nodes.  Linux made a similar change for the same reason, so increase to 32
to give us some extra breathing room as well.  The input and output arrays
are sized at 256, so much greater than the increase in the property array
size.


git-svn-id: https://svn.freebsd.org/base/head@332984 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f