forthloader: Remove "EFI boot environment message"
Contrary to what the message says, this is not only executed in an EFI
context- it provides functions for use in an EFI environment. I don't think
there's much reason to broadcast this fact when we haven't in the past, so
just remove it.

Reported by:	emaste (a while ago), cperciva


git-svn-id: https://svn.freebsd.org/base/head@332985 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f