MFC r333084:
pfctl: Don't break connections on skipped interfaces on reload

On reload we used to first flush everything, including the list of skipped
interfaces.  This can lead to termination of these connections if they send
packets before the new configuration is applied.

Note that this doesn't currently happen on 12 or 11, because of special EACCES
handling introduced in r315514. This special behaviour in tcp_output() may
change, hence the fix in pfctl.

PR:		214613
Submitted by:	Andreas Longwitz <longwitz at incore.de>


git-svn-id: https://svn.freebsd.org/base/stable/11@333181 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f