MFC r332426: allow ZFS pool to have temporary name for duration of current import
The change adds -t <name> option to zpool create and -t option to zpool
import in its form with an old name and a new name.  This allows to
import (or create) a pool under a name that's different from its real,
permanent name without affecting that name.  This is useful when working
with VM images or images of other physical systems if they happen to
have a ZFS pool with the same name as the host system.

Sponsored by:	Panzura (porting)


git-svn-id: https://svn.freebsd.org/base/stable/10@333196 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f