MFC r333115: libibumad/umad.c: In get_port, ignore sysctl get rate errors
This can cause ibpanic in ibstat when width is not set properly
as can occur when Ethernet port is connected to InfiniBand fabric.

ibpanic: [8167] main: stat of IB device 'mlx5_0' failed: m

With this change, Rate is displayed as 0 with ibstat for
this scenario.

Approved by:    hselasky (mentor), kib (mentor)
Sponsored by:   Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@333217 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f