arm64: rk: Add support for setting pll rate
Add support for setting pll rate. On RockChip SoC two kind of plls are
supported, integer mode and fractional mode.
The two modes are intended to support more frequencies for the core plls.
While here change the recalc method as it appears that the datasheet is
wrong on the calculation method.


git-svn-id: https://svn.freebsd.org/base/head@333315 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f